move SmallModalChildren to Modal folder
