+++ app/src/molecules/Modal/SmallModalChildren.tsx
@@ -8,8 +8,8 @@ import {
   BORDERS,
 } from '@opentrons/components'
 
-import { StyledText } from '../../../atoms/text'
-import { Modal } from '..'
+import { StyledText } from '../../atoms/text'
+import { Modal } from '.'
 
 interface SmallModalChildrenProps {
   handleCloseMaxPinsAlert: () => void
