+++ app/src/molecules/Modal/__tests__/SmallModalChildren.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { renderWithProviders } from '@opentrons/components'
-import { SmallModalChildren } from './SmallModalChildren'
+import { SmallModalChildren } from '../SmallModalChildren'
 
 const props = {
   header: 'header',
